refactor(js): split profile/main.js and stats/main.js into modules (#67)
profile/main.js (1068 -> 462 lines):
- components/my-hardware.js (225) - local hardware prefill, paste parsing
- components/systems.js (275) - server systems table, default toggle
- components/my-reports.js (196) - reports table, publish/unpublish/delete

stats/main.js (1077 -> 308 lines):
- utils.js (102) - constants, formatting helpers
- filters.js (184) - filter state, dropdown UI, URL sync
- charts.js (494) - all chart/sparkline/donut renderers

No behavior changes. All tests pass.
