+++ app.html
@@ -57,6 +57,6 @@
 <script src="js/lib/topbar.js?v=38d23fe3"></script>
 <script src="js/lib/toast.js?v=7bb23014"></script>
 <script src="js/lib/gh-gist.js?v=346e3ea7"></script>
-<script type="module" src="js/app/main.js?v=fdd905a5"></script>
+<script type="module" src="js/app/main.js?v=feeb014c"></script>
 </body>
 </html>
