+++ profile.html
@@ -287,6 +287,6 @@
 <script src="js/lib/supabase-client.js?v=04813aa1"></script>
 <script src="js/lib/topbar.js?v=38d23fe3"></script>
 <script src="js/lib/toast.js?v=7bb23014"></script>
-<script type="module" src="js/profile/main.js?v=267ea616"></script>
+<script type="module" src="js/profile/main.js?v=828c6869"></script>
 </body>
 </html>
