+++ stats.html
@@ -723,6 +723,6 @@ h2 {
 
 </div></div>
 
-<script type="module" src="js/stats/main.js?v=a77bbad8"></script>
+<script type="module" src="js/stats/main.js?v=0ea0efb1"></script>
 </body>
 </html>
